# back/openapi.yaml
openapi: 3.1.0
info:
  title: VLR Backend
  version: 0.1.0
paths:
  /api/v1/labs/{lab_id}/guide:
    get:
      tags:
        - Лабораторные работы
      summary: Методичка файлом
      description: "Отдаёт приложенный файл методички.\n\nФайл открывается обычной ссылкой, поэтому токен принимается и в\nзаголовке, и в параметре — иначе в WebView MAX скачивание не работает."
      operationId: lab_guide_api_v1_labs__lab_id__guide_get
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
            title: Lab Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema: {}
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/labs/:
    get:
      tags:
        - Лабораторные работы
      summary: Get Labs
      operationId: get_labs_api_v1_labs__get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LabResponse"
                type: array
                title: Response Get Labs Api V1 Labs  Get
      security:
        - HTTPBearer: []
    post:
      tags:
        - Лабораторные работы
      summary: Create Lab
      operationId: create_lab_api_v1_labs__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabCreate"
        required: true
      responses:
        201:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
  /api/v1/labs/{lab_id}:
    get:
      tags:
        - Лабораторные работы
      summary: Get Lab
      operationId: get_lab_api_v1_labs__lab_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
            title: Lab Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    patch:
      tags:
        - Лабораторные работы
      summary: Update Lab
      operationId: update_lab_api_v1_labs__lab_id__patch
      security:
        - HTTPBearer: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
            title: Lab Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabUpdate"
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/labs/{lab_id}/assign:
    post:
      tags:
        - Лабораторные работы
      summary: Выдать работу классам
      operationId: assign_lab_api_v1_labs__lab_id__assign_post
      security:
        - HTTPBearer: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
            title: Lab Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRequest"
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/v1/links/:
    get:
      tags:
        - Связи учитель — ученик
      summary: Мои связанные ученики
      operationId: list_links_api_v1_links__get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LinkResponse"
                type: array
                title: Response List Links Api V1 Links  Get
      security:
        - HTTPBearer: []
    post:
      tags:
        - Связи учитель — ученик
      summary: Связать ученика с учителем
      operationId: create_link_api_v1_links__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkCreate"
        required: true
      responses:
        201:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
...
components:
  schemas:
    AssignRequest:
      properties:
        classroom_ids:
          items:
            type: integer
          type: array
          title: Classroom Ids
      type: object
      title: AssignRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LabCreate:
      properties:
        title:
          type: string
          title: Title
        subject:
          anyOf:
            - type: string
            - type: null
          title: Subject
        summary:
          anyOf:
            - type: string
            - type: null
          title: Summary
        goal:
          anyOf:
            - type: string
            - type: null
          title: Goal
        guide:
          anyOf:
            - type: string
            - type: null
          title: Guide
        guide_file:
          anyOf:
            - type: string
            - type: null
          title: Guide File
        guide_file_name:
          anyOf:
            - type: string
            - type: null
          title: Guide File Name
        visibility:
          type: string
          enum:
            - public
            - private
          title: Visibility
          default: private
        math_model:
          additionalProperties: true
          type: object
          title: Math Model
        scene:
          anyOf:
            - additionalProperties: true
              type: object
            - type: null
          title: Scene
        journal_columns:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Journal Columns
        assigned_class_ids:
          items:
            type: string
          type: array
          title: Assigned Class Ids
        noise_percent:
          anyOf:
            - type: number
            - type: null
          title: Noise Percent
        max_rows:
          anyOf:
            - type: integer
            - type: null
          title: Max Rows
        is_published:
          type: boolean
          title: Is Published
          default: false
      type: object
      required:
        - title
        - math_model
      title: LabCreate
    LabResponse:
      properties:
        title:
          type: string
          title: Title
        subject:
          anyOf:
            - type: string
            - type: null
          title: Subject
        summary:
          anyOf:
            - type: string
            - type: null
          title: Summary
        goal:
          anyOf:
            - type: string
            - type: null
          title: Goal
        guide:
          anyOf:
            - type: string
            - type: null
          title: Guide
        guide_file:
          anyOf:
            - type: string
            - type: null
          title: Guide File
        guide_file_name:
          anyOf:
            - type: string
            - type: null
          title: Guide File Name
        visibility:
          type: string
          enum:
            - public
            - private
          title: Visibility
          default: private
        math_model:
          additionalProperties: true
          type: object
          title: Math Model
        scene:
          anyOf:
            - additionalProperties: true
              type: object
            - type: null
          title: Scene
        journal_columns:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Journal Columns
        assigned_class_ids:
          items:
            type: string
          type: array
          title: Assigned Class Ids
        noise_percent:
          anyOf:
            - type: number
            - type: null
          title: Noise Percent
        max_rows:
          anyOf:
            - type: integer
            - type: null
          title: Max Rows
        is_published:
          type: boolean
          title: Is Published
          default: false
        id:
          type: integer
          title: Id
        author_id:
          type: integer
          title: Author Id
        created_at:
          type: string
          format: "date-time"
          title: Created At
        author_name:
          anyOf:
            - type: string
            - type: null
          title: Author Name
      type: object
      required:
        - title
        - math_model
        - id
        - author_id
        - created_at
      title: LabResponse
    LabUpdate:
      properties:
        title:
          anyOf:
            - type: string
            - type: null
          title: Title
        subject:
          anyOf:
            - type: string
            - type: null
          title: Subject
        summary:
          anyOf:
            - type: string
            - type: null
          title: Summary
        goal:
          anyOf:
            - type: string
            - type: null
          title: Goal
        guide:
          anyOf:
            - type: string
            - type: null
          title: Guide
        guide_file:
          anyOf:
            - type: string
            - type: null
          title: Guide File
        guide_file_name:
          anyOf:
            - type: string
            - type: null
          title: Guide File Name
        visibility:
          anyOf:
            - type: string
              enum:
                - public
                - private
            - type: null
          title: Visibility
        math_model:
          anyOf:
            - additionalProperties: true
              type: object
            - type: null
          title: Math Model
        scene:
          anyOf:
            - additionalProperties: true
              type: object
            - type: null
          title: Scene
        journal_columns:
          anyOf:
            - items:
                additionalProperties: true
                type: object
              type: array
            - type: null
          title: Journal Columns
        assigned_class_ids:
          anyOf:
            - items:
                type: string
              type: array
            - type: null
          title: Assigned Class Ids
        noise_percent:
          anyOf:
            - type: number
            - type: null
          title: Noise Percent
        max_rows:
          anyOf:
            - type: integer
            - type: null
          title: Max Rows
        is_published:
          anyOf:
            - type: boolean
            - type: null
          title: Is Published
      type: object
      title: LabUpdate
    LinkCreate:
      properties:
        student_max_user_id:
          type: integer
          title: Student Max User Id
      type: object
      required:
        - student_max_user_id
      title: LinkCreate
    LinkResponse:
      properties:
        student_id:
          type: integer
          title: Student Id
        max_user_id:
          type: integer
          title: Max User Id
        full_name:
          type: string
          title: Full Name
        class_number:
          anyOf:
            - type: integer
            - type: null
          title: Class Number
        school:
          anyOf:
            - type: string
            - type: null
          title: School
      type: object
      required:
        - student_id
        - max_user_id
        - full_name
      title: LinkResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer